% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollnet.R
\name{plot_pollnet_particles}
\alias{plot_pollnet_particles}
\title{Synthetic plot of the particles' classification}
\usage{
plot_pollnet_particles(p = pollnet_particles(), plot_others = F)
}
\arguments{
\item{p}{Particles metadata (data.frame)}

\item{plot_others}{If TRUE, all classes are included}
}
\description{
Synthetic plot of the particles' classification
}
\examples{
plot_pollnet_particles(plot_others = FALSE)
plot_pollnet_particles(plot_others = TRUE)
}
