% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollnet.R
\name{pollnet_data_station}
\alias{pollnet_data_station}
\alias{pollnet_data}
\title{Get data}
\usage{
pollnet_data_station(part_id = NULL, from = NULL, to = NULL, stat_id,
  verbose = F)

pollnet_data(part_ids, from, to, ...)
}
\arguments{
\item{part_id}{Numeric code of the particle (pollen or spore)}

\item{from}{First day (string YYYY-MM-DD)}

\item{to}{Last day (string YYYY-MM-DD)}

\item{stat_id}{Numeric code of the station}

\item{verbose}{Logical. If TRUE, the URL is printed.}

\item{part_ids}{Vector of numeric codes of the particles (pollen and/or spore)}

\item{...}{Additional parameters passed to pollnet_data_station}
}
\value{
A data.frame with columns
\describe{
  \item{PART_SEQ}{progressive number of the particle}
  \item{PART_LEVEL}{hierarchy level}
  \item{PART_ID}{numeric code of the particle}
  \item{PART_NAME_L}{scientific name (Latin) of the particle}
  \item{PART_PARENT_NAME_L}{name of the group to which the particle is assigned}
  \item{REMA_CONCENTRATION}{average daily particle concentration}
  \item{REMA_DATE}{date}
  \item{STAT_ID}{numeric code assigned to the monitoring station}
  \item{STAT_CODE}{alphanumeric code of the monitoring station}
  \item{STAT_NAME_I}{Italian name of the monitoring station}
  \item{STAT_NAME_D}{German name of the monitoring station}
}
}
\description{
\code{pollnet_data_station} returns data of a single particle type, of a single monitoring station (max 5000 rows)

\code{pollnet_data} returns data of all the monitoring stations
}
\examples{
pollnet_data_station(part_id = 1352, from = "2018-01-01", to = "2018-01-10", stat_id = 84)
}
