% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollnet.R
\name{pollnet_overview}
\alias{pollnet_overview}
\title{Synthetic plot of the available pollen and spores}
\usage{
pollnet_overview(from = "2019-03-01", to = "2019-06-01")
}
\arguments{
\item{from}{First day (string YYYY-MM-DD)}

\item{to}{Last day (string YYYY-MM-DD)}
}
\description{
Synthetic plot of the available pollen and spores
}
\examples{
\dontrun{
p <- pollnet_overview(from="2019-03-01", to="2019-03-01")
print(p)#' ggsave(p, filename=paste0("overview_",from,"_",to,".pdf"), height=22,width=14)
}
}
