% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollnet.R
\name{pollnet_particles}
\alias{pollnet_particles}
\title{Get description of the monitored pollen and spores}
\usage{
pollnet_particles()
}
\value{
A data.frame with columns
\describe{
  \item{PART_ID}{numeric code of the particle}
  \item{PARENT_ID}{numeric code of the group to which the particle is assigned}
  \item{PARENT_NAME_L}{Latin name of the group to which the particle is assigned}
  \item{PART_CODE}{short code (string)}
  \item{PART_NAME_I}{Italian name}
  \item{PART_NAME_D}{German name}
  \item{PART_NAME_E}{English name}
  \item{PART_NAME_F}{French name}
  \item{PART_NAME_L}{Latin name}
  \item{PART_LOW}{lower limit for the concentration class "low"}
  \item{PART_MIDDLE}{lower limit for the concentration class "middle"}
  \item{PART_HIGH}{lower limit for the concentration class "high"}
}
}
\description{
Get description of the monitored pollen and spores
}
\examples{
summary(pollnet_particles())
}
