% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollnet.R
\name{pollnet_regions}
\alias{pollnet_regions}
\title{Get codes of Italian regions}
\usage{
pollnet_regions()
}
\value{
A data.frame with columns REGI_ID (integer ID), REGI_NAME_I (Italian name) and REGI_NAME_D (German name)
}
\description{
Get codes of Italian regions
}
\examples{
summary(pollnet_regions())
}
