% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollnet.R
\name{pollnet_stations}
\alias{pollnet_stations}
\title{Get metadata of the monitoring stations}
\usage{
pollnet_stations(regi_id = NULL)
}
\arguments{
\item{regi_id}{Numeric code of the region. If NULL, all regions are considered}
}
\value{
A data.frame with columns
\describe{
  \item{STAT_ID}{Numeric code}
  \item{STAT_CODE}{Short code (string)}
  \item{STAT_NAME_I}{Italian name}
  \item{STAT_NAME_D}{German name}
  \item{REGI_ID}{Numeric code of the region}
  \item{REGI_NAME_I}{Italian name of the region}
  \item{REGI_NAME_D}{German name of the region}
  \item{LATITUDE}{Latitude}
  \item{LONGITUDE}{Longitude}
}
}
\description{
Get metadata of the monitoring stations
}
\examples{
summary(pollnet_stations(regi_id=13))
}
